refactor: searchBar with hooks
